# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "retrieval-rest-server",
    "version": "0.1.0"
  },
  "paths": {
    "/api/upsert_document": {
      "post": {
        "summary": "Api Upsert",
        "description": "Upsert a document into the store.",
        "operationId": "api_upsert_api_upsert_document_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/index_path": {
      "post": {
        "summary": "Api Index Path",
        "description": "Index a filesystem path into the retrieval store.",
        "operationId": "api_index_path_api_index_path_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndexPathReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/search": {
      "post": {
        "summary": "Api Search",
        "description": "Search the retrieval store.",
        "operationId": "api_search_api_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/grounded_answer": {
      "post": {
        "summary": "Api Grounded Answer",
        "description": "Return a grounded answer using vector search + synthesis.",
        "operationId": "api_grounded_answer_api_grounded_answer_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroundedAnswerReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rerank": {
      "post": {
        "summary": "Api Rerank",
        "description": "Rerank provided passages for a query.",
        "operationId": "api_rerank_api_rerank_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RerankReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/verify_grounding": {
      "post": {
        "summary": "Api Verify",
        "description": "Verify grounding against store using citation URIs.",
        "operationId": "api_verify_api_verify_grounding_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/documents/delete": {
      "post": {
        "summary": "Api Documents Delete",
        "description": "Delete documents by URI from the store and rebuild index.",
        "operationId": "api_documents_delete_api_documents_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteDocsReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/extract": {
      "post": {
        "summary": "Api Extract",
        "description": "Extract text from an uploaded file (in-memory, no storage).",
        "operationId": "api_extract_api_extract_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_extract_api_extract_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_api_extract_api_extract_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_api_extract_api_extract_post"
      },
...
      "DeleteDocsReq": {
        "properties": {
          "uris": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Uris"
          }
        },
        "type": "object",
        "required": [
          "uris"
        ],
        "title": "DeleteDocsReq",
        "description": "Request model for deleting documents by URI."
      },
...
      "GroundedAnswerReq": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          },
          "k": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "K",
            "default": 3
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "max_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Tokens"
          },
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "GroundedAnswerReq",
        "description": "Request model for grounded answer generation."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IndexPathReq": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "glob": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Glob",
            "default": "**/*.txt"
          }
        },
        "type": "object",
        "required": [
          "path"
        ],
        "title": "IndexPathReq",
        "description": "Request model for indexing a filesystem path."
      },
...
      "RerankReq": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "passages": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Passages"
          },
          "top_k": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top K"
          }
        },
        "type": "object",
        "required": [
          "query",
          "passages"
        ],
        "title": "RerankReq",
        "description": "Request model for reranking passages."
      },
      "SearchReq": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "async": {
            "type": "boolean",
            "title": "Async",
            "default": false
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Seconds",
            "default": 300
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "max_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Tokens"
          },
          "top_k": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top K"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SearchReq",
        "description": "Request model for performing a search."
      },
      "UpsertReq": {
        "properties": {
          "uri": {
            "type": "string",
            "title": "Uri"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "binary_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Binary Base64"
          }
        },
        "type": "object",
        "required": [
          "uri"
        ],
        "title": "UpsertReq",
        "description": "Request model for upserting a document."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerifyReq": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          },
          "draft_answer": {
            "type": "string",
            "title": "Draft Answer"
          },
          "citations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Citations"
          }
        },
        "type": "object",
        "required": [
          "question",
          "draft_answer",
          "citations"
        ],
        "title": "VerifyReq",
        "description": "Request model for grounding verification (citations)."
      },
...
    }
  }
}